refactor(api): restructure demarches-simplifiees service and utils
Split the service into client/functions/index, move the base URL, slug map and createDemarcheNumeriqueDossier helper to utils/demarches-simplifiees, rename the job to demarches-simplifiees-apply-import, and replace the paginated while(true) with a hasNextPage flag.
