+++ api/src/config.ts
@@ -100,8 +100,6 @@ export const PILOTY_BASE_URL = process.env.PILOTY_BASE_URL || "https://sandbox-a
 export const LETUDIANT_PILOTY_TOKEN = process.env.LETUDIANT_PILOTY_TOKEN || "";
 
 // Démarches Simplifiées (API GraphQL)
-// Instance qui héberge la démarche (par défaut demarche.numerique.gouv.fr, l'instance DINUM).
-export const DEMARCHES_SIMPLIFIEES_BASE_URL = process.env.DEMARCHES_SIMPLIFIEES_BASE_URL || "https://demarche.numerique.gouv.fr";
 export const DEMARCHES_SIMPLIFIEES_TOKEN = process.env.DEMARCHES_SIMPLIFIEES_TOKEN;
 
 // Typesense
