+++ api/src/services/demarches-simplifiees/index.ts
@@ -1,0 +1,4 @@
+import { query } from "./client";
+import { createDossier, getAllDossiers, getDemarche, getDemarcheNumberBySlug, getDossier } from "./functions";
+
+export default { query, getDemarche, getDossier, getAllDossiers, createDossier, getDemarcheNumberBySlug };
