+++ api/src/utils/redirect.ts
@@ -1,56 +1,7 @@
 import { JVA_URL, PUBLISHER_IDS } from "@/config";
 
-import demarchesSimplifiees, { DEMARCHE_MAP } from "@/services/demarches-simplifiees";
 import { statBotService } from "@/services/stat-bot";
 import { statEventService } from "@/services/stat-event";
-
-const DEMARCHE_NUMERIQUE_HOST = "demarche.numerique.gouv.fr";
-
-// Si l'URL de candidature pointe vers demarche.numerique.gouv.fr, on crée un dossier prérempli côté
-// Démarches Simplifiées. On renvoie son numéro (à stocker dans customAttributes) et son URL (cible de
-// redirection à la place de l'applicationUrl générique). Le numéro de démarche est résolu depuis le
-// slug de l'URL : d'abord via DEMARCHE_MAP, sinon via l'API (getDemarcheNumberBySlug).
-export const createDemarcheNumeriqueDossier = async (applicationUrl: string | null | undefined): Promise<{ number: number; url: string } | null> => {
-  if (!applicationUrl) {
-    return null;
-  }
-
-  let parsedUrl: URL;
-  try {
-    parsedUrl = new URL(applicationUrl);
-  } catch (error) {
-    return null;
-  }
-  if (parsedUrl.hostname !== DEMARCHE_NUMERIQUE_HOST) {
-    return null;
-  }
-
-  const slug = parsedUrl.pathname.match(/\/commencer\/([^/?#]+)/)?.[1];
-  if (!slug) {
-    return null;
-  }
-
-  let demarcheNumber = DEMARCHE_MAP[slug];
-  if (!demarcheNumber) {
-    const found = await demarchesSimplifiees.getDemarcheNumberBySlug(slug);
-    if (!found.ok) {
-      return null;
-    }
-    demarcheNumber = found.data;
-  }
-
-  const result = await demarchesSimplifiees.createDossier(demarcheNumber);
-  if (!result.ok) {
-    return null;
-  }
-
-  const number = result.data?.dossier_number ?? null;
-  const url = result.data?.dossier_url ?? null;
-  if (!number || !url) {
-    return null;
-  }
-  return { number, url };
-};
 
 const ensureUrlProtocol = (href: string) => {
   if (href.indexOf("http://") === 0 || href.indexOf("https://") === 0) {
